pyproject.toml
[tool.poetry]
name = "base-user"
version = "1.0"
description = "Descrição do seu projeto"
authors = ["Isis Araujo <isis.araujo@outlook.com>"]
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
Django = "^4.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
django-phonenumber-field = "^7.2.0"
drf-yasg = "^1.21.7"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
phonenumbers = "^8.13.0"
django-admin-interface = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.7.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
black = "^23.0"
flake8 = "^6.0"
pre-commit = "^4.0.1"
factory-boy = "^3.3.0"
faker = "^19.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=. --cov-report=html"
